Circular Reference fixed Index not running Main() fixed. getAxis, getButton on keyboard/Controller fixed
removed unused/non existing imports

moved mobileAndTabletCheck() to Utils
