+++ index.html
@@ -12,6 +12,6 @@
 </head>
 <body>
     <canvas id="mainCanvas"></canvas>
-    <script type="module" src="./js/engine/run.js"></script>
+    <script type="module" src="./js/Main.js"></script>
 </body>
 </html>
